drivers: ethernet: microchip: move into own dir
move microchip ethernet drivers into own directory.
