+++ drivers/ethernet/CMakeLists.txt
@@ -13,6 +13,7 @@ endif()
 
 add_subdirectory(dsa)
 add_subdirectory(phy)
+add_subdirectory(microchip)
 add_subdirectory(nxp)
 add_subdirectory(dwc_mac)
 add_subdirectory(dwc_xgmac)
@@ -28,20 +29,15 @@ zephyr_library_sources_ifdef(CONFIG_ETH_CYCLONEV eth_cyclonev.c)
 zephyr_library_sources_ifdef(CONFIG_ETH_DM9051 eth_dm9051.c)
 zephyr_library_sources_ifdef(CONFIG_ETH_DRIVER_RAW_MODE eth_raw.c)
 zephyr_library_sources_ifdef(CONFIG_ETH_E1000 eth_e1000.c)
-zephyr_library_sources_ifdef(CONFIG_ETH_ENC28J60 eth_enc28j60.c)
-zephyr_library_sources_ifdef(CONFIG_ETH_ENC424J600 eth_enc424j600.c)
 zephyr_library_sources_ifdef(CONFIG_ETH_ESP32 eth_esp32.c)
 zephyr_library_sources_ifdef(CONFIG_ETH_GECKO eth_gecko.c)
 zephyr_library_sources_ifdef(CONFIG_ETH_GECKO phy_gecko.c)
 zephyr_library_sources_ifdef(CONFIG_ETH_IVSHMEM eth_ivshmem.c eth_ivshmem_queue.c)
-zephyr_library_sources_ifdef(CONFIG_ETH_LAN865X eth_lan865x.c oa_tc6.c)
-zephyr_library_sources_ifdef(CONFIG_ETH_LAN9250 eth_lan9250.c)
+zephyr_library_sources_ifdef(CONFIG_ETH_LAN865X oa_tc6.c)
 zephyr_library_sources_ifdef(CONFIG_ETH_LITEX_LITEETH eth_litex_liteeth.c)
-zephyr_library_sources_ifdef(CONFIG_ETH_MCHP_GMAC_G1 eth_mchp_gmac_g1.c)
 zephyr_library_sources_ifdef(CONFIG_ETH_NUMAKER eth_numaker.c)
 zephyr_library_sources_ifdef(CONFIG_ETH_RENESAS_RA eth_renesas_ra.c)
 zephyr_library_sources_ifdef(CONFIG_ETH_RENESAS_RA_RMAC eth_renesas_ra_rmac.c)
-zephyr_library_sources_ifdef(CONFIG_ETH_SAM_GMAC eth_sam_gmac.c)
 zephyr_library_sources_ifdef(CONFIG_ETH_SMSC911X eth_smsc911x.c)
 zephyr_library_sources_ifdef(CONFIG_ETH_SMSC91X eth_smsc91x.c)
 zephyr_library_sources_ifdef(CONFIG_ETH_STELLARIS eth_stellaris.c)
