+++ drivers/ethernet/Kconfig
@@ -55,20 +55,14 @@ source "drivers/ethernet/Kconfig.bcmgenet"
 source "drivers/ethernet/Kconfig.cyclonev"
 source "drivers/ethernet/Kconfig.dm9051"
 source "drivers/ethernet/Kconfig.e1000"
-source "drivers/ethernet/Kconfig.enc28j60"
-source "drivers/ethernet/Kconfig.enc424j600"
 source "drivers/ethernet/Kconfig.esp32"
 source "drivers/ethernet/Kconfig.gecko"
 source "drivers/ethernet/Kconfig.ivshmem"
-source "drivers/ethernet/Kconfig.lan865x"
-source "drivers/ethernet/Kconfig.lan9250"
 source "drivers/ethernet/Kconfig.litex"
-source "drivers/ethernet/Kconfig.mchp"
 source "drivers/ethernet/Kconfig.native_tap"
 source "drivers/ethernet/Kconfig.numaker"
 source "drivers/ethernet/Kconfig.renesas_ra"
 source "drivers/ethernet/Kconfig.renesas_ra_rmac"
-source "drivers/ethernet/Kconfig.sam_gmac"
 source "drivers/ethernet/Kconfig.smsc911x"
 source "drivers/ethernet/Kconfig.smsc91x"
 source "drivers/ethernet/Kconfig.stellaris"
@@ -87,6 +81,7 @@ source "drivers/ethernet/dsa/Kconfig"
 source "drivers/ethernet/dwc_mac/Kconfig"
 source "drivers/ethernet/dwc_xgmac/Kconfig"
 source "drivers/ethernet/intel/Kconfig"
+source "drivers/ethernet/microchip/Kconfig"
 source "drivers/ethernet/nxp/Kconfig"
 source "drivers/ethernet/offload/Kconfig"
 source "drivers/ethernet/phy/Kconfig"
