+++ drivers/ethernet/mdio/mdio_lan865x.c
@@ -15,7 +15,7 @@ LOG_MODULE_REGISTER(mdio_lan865x, CONFIG_MDIO_LOG_LEVEL);
 #include <zephyr/kernel.h>
 #include <zephyr/drivers/mdio.h>
 
-#include "../eth_lan865x_priv.h"
+#include "../microchip/eth_lan865x_priv.h"
 
 struct mdio_lan865x_config {
 	const struct device *dev;
